controllervm controller
Creates controller VMs on CNV base on the controlplane CR. The
base image, network-attachment and virtual machine is created
as unstructured. A secret is created for the cloudinit userdata
using a template.
The controlplane controller create a secret for the deployment
user cloud-admin. The cloud-admin user and the pubkey from the
secret gets added via cloudinit.