Add ProvisionServerName to BaremetalSet CR
This adds a new ProvisionServerName to BaremetalSet that
can be used as an alternate to RhelImageUrl.

Also updates BaremetalSet so that either RhelImageUrl or
ProvisionServerName needs to be set.

Pre-creating a ProvisionServer is desirable if you plan on sharing
the same image with multiple BaremetalSets. Also, pre-creating a
provisionserver should speed up future demos.