Fix OSPNetwork due to apiextensions.k8s.io/v1 change
With switch to apiextensions.k8s.io/v1 pruning is enabled, by default [1].
This moves toe use NodeNetworkConfigurationPolicySpec as it is what is
used/created anyway. In later probably want to move from unstructured to
create the NodeNetworkConfigurationPolicy

Also moves the OSPNetwork to be global OpenStackControlPlaneSpec paremeter
instead of per OpenStackVirtualMachineRole as it is the same for all virtual
machines.

[1] https://kubernetes.io/blog/2019/06/20/crd-structural-schema/#pruning-don-t-preserve-unknown-fields