Support tripleo network isolation using vlan
To attach the controller VM to the different VLANs, for each
VLAN a network-attachment-definition gets created, which was
done via the refactoring of the OSPnetwork in [1].

Adds optional OpenStackClientNetworks list parameter to the
OpenStackControlPlaneSpec CRD to allow customization of the
networks the OpenStackClient pod gets attached. Default is
ctlplane and external. For debugging it can be usefull to be
able to attach a openstackclient pod to any of the created
networks.

[1] https://github.com/openstack-k8s-operators/osp-director-operator/pull/180