Switch away from bindata for vm resources
This switch to use the schema for NetworkConfigurationPolicy,
NetworkAttachmentDefinition and VirtualMachine instead of the
previous used bindata templates.

For NetworkConfigurationPolicy and NetworkAttachmentDefinition
to use dynamic client + ssa as CreateOrUpdate can not be used
because NetworkConfigurationPolicy is non namespaced.

This just changes the current implementation to use the above.
This is reqired to make it easier to enabled support for
network isolation where a dynamic list of
NetworkAttachmentDefinition needs to be created depending on
the used OSP network list.

With this change it is now also possible to modify vm resources
by changing the CR parameters, e.g. it is now possible to change
configuration parameters, like CPU/memory/... for the controllers.
When changed the controller pod needs to be recreated.
(virtctl stop/start controller-0)