Support deploy ComputeHCI
Adds support for Compute nodes also be Ceph OSP nodes.

Also splits the tripleo-deploy.sh to have -r and -p option for
the two phases:
-r use ephemeral heat to render the ansible playbooks
-p run ansible playbooks

With this we can control which part of the deployment process
to run.

Note: right now skip-tags opendev-validation-ceph needs to be
added to the deployment command if ceph gets deployed. Ceph
validations checks if lvm2 package is installed. The check
verifies if the package is in any host of the ansible facts, which
includes the "undercloud". Since the lvm2 package is not installed
in the openstackclient container image this fails.