IP handling during scaling
Scaling down has the following workflow:
1) annotate VM or BMH which should be removed
2) decrease the resource count of OSBMset/OSCtlplane

Right now the IP reservation entry won't be set as deleted.
As a result, we'll still have the deleted hosts in the
rendered predictible IPs list of the tripleo environment
file.

This change introduce the following new behavior:
* annotate VM or BMH with a unified delete annotation
  osp-director.openstack.org/delete-host=true
* the annotation status of the VM/BMH gets reflected in the
  resource status via "annotatedForDeletion": false/true
* decrease the resource count of OSBMset/OSCtlplane to
  remove the annotated VM/BMH
* when node gets deleted it is deleted from
  - the resource CR status (OSVMSet/OSBMset)
  - the node gets removed from the ipset status
  - the osnet status entry get set as deleted
  -> as a result the rendered tripleo IPs list for the role
     will have a deleted entry
* right now the IPreservation is only marked as deleted,
  if a new node gets scaled in it will get the same hostname
  and IP address.

The osnet status is now a map[string]OpenStackNetRoleStatus,
where the OpenStackNetRoleStatus
- has the flag if all entries from that role should be added
  to the tripleo predictible IPs list (AddToPredictableIPs)
- list of IPReservations

With this structur, if the corresponding role CR gets deleted
the IP reservations can all be deleted at once via the added
finalizer in the OSIPset controller.

Additional changes:
- use common LabelSelectors for controllers
- simplifies the IPSet status.network status to not contain
  AttachConfiguration specifics