Remove rhos-release and document rhsm usage
This removes the installation of rhos-release package vi cloud-init
and documents the procedure how to register the overcloud nodes
via ansible to RHSM/Satellite.

Also moves the checkout of the accepted rendered playbooks to the
tripleo-deploy.sh -a action that the updated inventory.yaml file
is available to be used to register the overcloud nodes before
running the playbooks with -p.

Depends-On: https://github.com/openstack-k8s-operators/osp-director-dev-tools/pull/146