Change nncp to be network independent
At the moment we create for each network a nncp and a
network-attachment-definition corrensponding to the network name.
This results in duplicate nncp for the same bridge if more then
one network use the same bridge which can result in config race.

This patch changes the nncp creation to create the nncp for the
bridge instead of the network. With this we only create a single
nncp for required bridges.

Notes:
- this change only creates/updates the nncp for the desired
  config in the osnet cr. It does not verify if the config is
  different from what the current nncp might be. In case two
  networks use the same bridge name, but different interfaces,
  resulting from a typo/missconfig. The nncp gets updated and
  might result in an update loop.
- since the osnet controller waits for the nncp to be created
  before it creates the network-attachment-definition, the check
  for the nncp in the osvmset/osclient controller were not
  required and are now removed.
- at the moment there is no finalizer no finalizer on the nncp.
  It is not an issue right now as we do not cleanup the nncp on
  osnet cr delete. When we add nncp cleanup we need to make sure
  to add finalizers on the nncp for each network that the nncp
  only gets deleted when all networks are gone.