fix vip creation
At the moment we run over the virtualMachineRoles and create for
the network list of the role vips for the network. If there are
multiple roles with different network list, last virtualMachineRoles
network list updates the ipset and VIPs won't get created.

This changes to create a uniq list of all networks for all
virtualMachineRoles and creates VIPs for this uniq list.