Add finalizer to osephemeralheat
When the configmap in osplaybookgenerator changes, the
ephemeralheat gets recreated, but there is no finalizer used
yet to wait that all resource got deleted. When the resource
get not cleaned up correct, the password in the still running
mariadb pod and the heat pods can be different and the
playbookgeneration is broken.

This introduce a finalizer on osephemeralheat to cleanup all
resources properly when the osplaybookgenerator triggers a
recreate of the ephemeral heat stack.