Make sure OpenStackClient creates IPset with uniq Role
If multiple OpenStackClient CRs get created the created right now
the IPset with the same Role parameter. This results in the
OpenStackClient pods to use the same IP addresses because the
OSNet reservations are grouped by the Roles.
This change adds the OpenStackClient CR name to the OpenStackClient
Role name const to make sure the Role name passed to create the
IPSet CR is uniq.