Support rendering required parameter files for osp16 and 17
The osctlplane controller checks the tripleoclient image url and
gets the OSP version from the image tag. Expected to have either
16.2 or 17.0. Everything after that tag prefix get ignored.

The OSP version gets recorded in the CR status of the osctlplane.
Other controllers which require this information can get the OSP
version from there. Right now this is the osipset and
osplaybookgeneratior.

Also updates create-playbooks.sh to match the inventory file name
and rendered playbook location to match what is done in 17.