Add script to help stage downstream rebases
This script helps push rebase commits downstream
and rebuilds the associated bundle container files
for that commit.

There are some definate release automation improvements we
might could make based on the logic here. Specifically
with regards to the dependency between the bundle image
and the final image:tag of the operator image.