Do not set owner on fencingData secret
As there can be multiple vm roles, where multiple vmsets get
created, it is not possible to set the owner to the fencing
secrets as this will otherwise result in an error to set the
ownershit.
This changes the VMset controller to return an err for the
result of generateNamespaceFencingData and reconcile instead of
nil and also skip ownership set on the fencing secret.