Add missing attach method condition
Attach method identification was hard coded for testing.
Now the attach method gets properly indentified from the
desired state of the the osnetatt config. Also the method
and the bridge name gets added to the osnetatt status for
easy use by the controllers without the need to analyse the
desired state config every time.