[OSPK8-416] reduce repeated condition message logging
Right now on every reconcile messages get logged, even if there
was no action.

To make logging better readable/ easier to follow, this change
logs the condition messages in the defer function where we check if
the condition changed. This should be used where we return from the
reconcile loop, and only in cases where this is possible log explicitely
at that place.