[OSPK8-422] generate ctlplane Ansible inventory
This patch creates an Ansible inventory of the cltplane
IPs within the openstackclient pod. The intent is for this
to be used for RH node registration instead of relying
on the ConfigGenerator/TripleO inventory to be generated
which doesn't actually exist at the time we need it with
the new workflow.