[OSPK8-421] Move OsMacAddr mgmt to osnetcfg
This moves the osmac reservations create to the osnetcfg controller
and updates the osmacaddr spec with new/changed reservations. For
this the osvmset and osbmset controller adds a label to its CR with
the osnetcfg CR name. This allows the osnetcfg to watch osbms and
osvmset CRs and reconciles on changes. It always recreates the full
list of reservations. Therefore no special handling is required
to cleanup reservations. If a node is gone from the osbms/osvmset
status also the mac reservation is gone.

The status of the osmacaddr CR still holds the same flattened
reservation list. This is still the source to render the ooo
parameter file.

This also changes the osbms to use the same common HostStatus type
to be able to use a common function to check the created osbms
and osvmset nodes. Right now required parameters from osbms got
merged into the existing HostStatus struct. In a follow up this
should be cleaned up (CtlplaneIP).