openstackclient pod dynamic Ansible inventory (#487)
* openstackclient pod dynamic Ansible inventory

This patch implements a dynamic Ansible inventory script
for the openstackclient pod. This inventory queries
the openstacknets API and parses out the relevant
host IPs which are then formatted in json for Ansible.

Docs are updated to recommend the new inventory is used
instead of having to rely on the static playbook generated
inventory we were using.

The motivation for this patch was to eliminate the dependency
on playbook generation to complete before node registration
could be accomplished, thus decoupling the two mechanism. This
dependency was a blocker for further deployment workflow changes.

* Typos

* No need to chmod in init.sh

* Rework to support roles

* Add ansible vars

* Updates

* pook

* Drop duplicate RHSM register details.

* script comments