[OSPK8-431] Migrate osipset functionality to osnetcfg
This migrates the IP reservation functionality from osipset to
osmacaddr controller. The osmacaddr controller watches osctlplane
osvmset, osclient and osbmset resources and creates IP reservations
if needed. The reservations get stored on the osnet spec, like before.
The ip set information is stored on the status of the osnetcfg CR.

If a node gets deleted, the default is to persist the reservations
of a node if the full role isn't deleted (like before). With this
change this can be controlled by the osnetcfg Spec.PreserveReservations
parameter, which controlls both IP and MAC address reservations.

In a follow up static IP reservation will be added.