[OSPK8-327] Allow static ip reservation
This allows todo static IP reservation for nodes by adding a
node entry to spec.reservations, like:

  reservations:
    controller-0:
      ipReservations:
        ctlplane: 192.168.25.40
      macReservations: {}
    controller-1:
      ipReservations:
        ctlplane: 192.168.25.41
      macReservations: {}
    controller-2:
      ipReservations:
        ctlplane: 192.168.25.42
      macReservations: {}
    controlplane:
      ipReservations:
        ctlplane: 192.168.25.10
      macReservations: {}
    openstackclient-0:
      ipReservations:
        ctlplane: 192.168.25.251
      macReservations: {}

Where node name is RoleName lower + index. For the VIPs of the control
plane use controlplane and openstackclient-0 for the openstackclient
pod.

Via webhook it gets verified if
* there are duplicates in the reservatios list to prevent that static
IPs get used for multiple hosts. In a follow up it should also get
checked if there are duplicates on update from static IPs and dynamic
created.
* if the IP address from static reservations have a valid IPv4/6
format.