[OSPK8-433] Allow to set StorageAccessMode and StorageVolumeMode for VMset
This patch allows to set the StorageAccessMode and/or StorageVolumeMode
for the VirtualMachineRole disk. StorageAccessMode defaults to
ReadWriteMany as this is required to be able to live migrate
an instance, but could be set to ReadWriteOnce if wanted.
StorageVolumeMode defaults to Filesystem, but when OpenShift
Virtualization with OpenShift Container Platform Container Storage
is used, Block should be used.

(cherry picked from commit c53e5504a7d943355b111b09cca92ec79f6d75d7)