Fix race in IP address generation due to different sources used
IP reservation flow is:
* osnetcfg creates IPs for new nodes
* osnetcfg updates osnet spec to new reservations
* osnet controller updates osnet status with reservations from spec

Right now GetAllIPReservations() use the osnet.spec.reservations
provided information to generate a list of all reserved IPs, while
if the osnetcfg controller used the osnet.status to check if a node
already had an IP.

If the osnetcfg controller reconciles while the osnet has not yet
finished to update the osnet.status from the new osnet.spec, it
can happen that for new nodes new IPs get created in the second
reconilation run:

first IP create for compute-0
2022-02-02T15:01:21.875Z        INFO    controllers.OpenStackNetConfig  BOOOO 2 reservations [{172.17.0.40 compute-0 false false}]
2022-02-02T15:01:21.875Z        INFO    controllers.OpenStackNetConfig  BOOOO 3 reservations [{172.17.0.40 compute-0 false false} {172.17.0.11 compute-1 false false}]
2022-02-02T15:01:21.875Z        INFO    controllers.OpenStackNetConfig  Created new reservation for host compute-1 in network internal_api with IP 172.17.0.11/24       {"ObjectType": "*v1beta1.OpenStackNetConfig", "ObjectNamespace": "openstack", "ObjectName": "openstacknetconfig"}

second IP create for compute-0
2022-02-02T15:01:22.035Z        INFO    controllers.OpenStackNetConfig  BOOOO 3 reservations [{172.17.0.40 compute-0 false false} {172.17.0.12 compute-1 false false}]
2022-02-02T15:01:22.035Z        INFO    controllers.OpenStackNetConfig  Created new reservation for host compute-1 in network internal_api with IP 172.17.0.12/24       {"ObjectType": "*v1beta1.OpenStackNetConfig", "ObjectNamespace": "openstack", "ObjectName": "openstacknetconfig"}
2022-02-02T15:01:22.035Z        INFO    controllers.OpenStackNetConfig  BOOOO 4 RoleReservations [{172.17.0.40 compute-0 false false} {172.17.0.11 compute-1 false false}]
2022-02-02T15:01:22.035Z        INFO    controllers.OpenStackNetConfig  BOOOO 5 RoleReservations {[{172.17.0.40 compute-0 false false} {172.17.0.11 compute-1 false false}] true}

While this is not really an issue as the first IP will be never used,
it results in having holes in the IP range which then get used for
later hosts.

This patch make sure that both situations (GetAllIPReservations
and verify if a host already has a reservation) use the same source,
osnet.status.