Fix adding osnetcfg and osnet labels on CR
Right now AddOSNetConfigRefLabel and AddOSNetNameLowerLabels
updated the object when added the required labels to the metadata.
Since the generic way of the functions only updated the object
labels at that point of the reconcilation other object information,
like status got lost.
This changes the two functions to merge the required labels into
the object labels and returns them so that they can be set on the
CR object instead of writing the object in the specific label
functions.

(cherry picked from commit 583bac834c17cbcf442dac8d7947fa92eeb0d4d6)