Fix duplicate ip assignmnent
Already assigned IPs right now are only considered as reserved when
they got added to the osnet status. If IPs for multiple roles get
created in the same run, new reservations are only added to the
osnet spec and not yet synced to the osnet status. Because of that
it might happen that IPs assigned/created for different get assigned
to nodes from another role.

This fix make sure that also new IP reservations added to the osnet
spec and not yet synced to the osnet status are considered as
reserved and therefore are not a candidate for new ip assignments.

(cherry picked from commit f5b611b5deab0877171147fb8d7ca95123e5220a)