Support interrupting a deploy job
The tripleo-deploy.sh oc exec shell creates a new process group. Can use
this to terminate the deployment.

Set terminateGracePeriod to 60s for the deploy job container.
Add a signal handler goroutine to the deploy job that exec a new
tripleo-deploy-term.sh script on the openstackclient.
This script will send a SIGTERM to the process group and, if it has not
not terminated after 30s, send a SIGKILL.