[OSPK8-447] fix network routes for IPv6 multi subnet
The routes parameter of the subnet in `network_data.yaml`, rendered
by the operator, use `routes` instead of the expected `routes_ipv6`.

Also the routes parameter rendered in environments/network-environment.yaml
which are named <netName>Routes get usually set on the neutron network
host_routes property.

~~~
  {{network.name}}Routes:
      Routes for the {{network.name_lower}} network traffic.
      JSON route e.g. [{'destination':'10.0.0.0/16', 'nexthop':'10.0.0.1'}]
      Routes are added to the host_routes property on the subnet in neutron
      when the network and subnet is created.
~~~

Since there is no neutron we add the {{network.name}}Routes to the nic template
and concat the two lists:

~~~
  routes:
    list_concat_unique:
      - get_param: {{ $net.Name }}Routes
      - get_param: {{ $net.Name }}InterfaceRoutes
~~~