Fix check for upstream/downstream version for AdditionalServiceVIPs
For wallaby/OSP17.0 AdditionalServiceVIPs are required for Redis
and OVN DB. Right now there is only a check for the 17.0 version
and not upstream wallaby [1].
This changes to use the GetOSPVersion [2] function to map the
provided/defaulted OSPVersion.

Also adds version check to the webhook to validate a supported
version is provided in the CR if any.

[1] https://github.com/openstack-k8s-operators/osp-director-operator/blob/master/api/v1beta1/openstackcontrolplane_webhook.go#L130
[2] https://pkg.go.dev/github.com/openstack-k8s-operators/osp-director-operator/api/v1beta1#GetOSPVersion