Ensure correct images are used for provision server
The provision server should always use the images specified in the
bundle CSV.
With the current approach of setting images in the spec and defaulting
in the webhook, users are likely to overlook updating the CRs after
updating the operator.
This could lead to mis-matched operator and agent images being used
together, or missing out on critial security fixes in the apache image.

This removes the images from the provision server spec completely and
ensures the correct images are always used in the Reconcile function.