Handle DeployIdentifier
DeployIdentifier should be unique to every deploy invocation. As we
generate playbooks and re-use them we cannot do this via a heat
parameter. Instead set this to a marker string when generating playbooks
that can be replaced later at deploy time.

Also add a deploy CR param to skip setting this, as it can be used to
speed up scale-up deployments.