Replicate 16.2 parameter merge behaviour
In a regular OSP16 deployment, parameters are merged by heat-client if
they are a map type.
This does not happen automatically in OSP17 regular deployments and
Director operator (for both 16.x and 17.x) which use ephemeral heat.
The merge strategy must be set explicitly in each environment file.

To workaround this, and remain compatible with existing 16.x
environment files and 3rd party integrations, automatically inject the
merge strategy for all map type parameters if not explicitly set.