Replicate StackAction behaviour for 16.2
Pacemaker restart, Octavia, and tripleo-network-config rely on the
stack_action (CREATE or UPDATE) ansible fact or hieradata.
To ensure this works correctly with director operator's ephemeral
heat we need to replicate the tripleoclient logic.
As we do not have the existance of a heat stack to rely on we need a
different approach.
Instead the existance of /var/lib/tripleo-config on any overcloud node
is used to determine the appropriate stack_action setting. This is
passed on the ansible-playbook command line to override the default
stack_action in global_vars (which is always CREATE). The resulting
hieradata is also updated.