[OSPK8-430] introduce backup v1beta2 api
The backup api always need to be on the latest api version
to be able to backup the latest version.

Note: when there is a api change, old backups can not be
restored due to the missing conversion webhook. Backup
containing CRs with older api versions have to be updated
manually before restore it.