[OSPK8-398] anti-affinity placement of controller VMs
Set default affinity rules that two pods of the same role should
not run if possible on the same worker node. With this it is still
possible to live migrate VMs and they run on the same node, but
on the next scheduling request (live migration) they get distributed
again.