Fix VM role nic template
Since the switch to use role template for the virtual machines, network
configuration is broken when using multiple VM roles which don't use
the same networks. Basically at the moment the roles nic template still
renders the networks for the roles which results
nic-template.role.j2.yaml has fixed network parameter list of the last
vm role instead of the jinja templating. Therefore process-templates.py
does not create the correct nic template for the specified networks
for the role.

This changes the net-config-multi-nic.role.j2.yaml to be almost raw
ooo roles nic template. The only go templated is to set the
ClusterServiceIP for the pod interface to be able to connect to the
OCP cluster for fencing.

Note: templated information needs to be translated from
{{network.name_lower}} to using golang raw string constant like
{{`{{network.name_lower}}`}}.