[OSPK8-600] Don't delete role reservations with PreserveReservations true
In master of the osp-director-operator OSIPSet got introduced.
During the update from 1.2.x to latest the OSIPSets get created.
The OSNetConfig depends on the OSIPSet to identify which
reservations could be deleted. If an OSIPSet is gone the OSNetConfig
expects that the reservations for the whole role could be delted.

This conflicts during the update as there is no OSIPSet and therefore
the OSNetConfig will delete all reservations, only the nodes which
have static reservations are save to keep their current IPs.

This changes the OSNetConfig behaviour to not delete reservations
for deleted roles if Spec.PreserveReservations is set to true,
which is the default behaviour.