Move EnableFencing to OpenStackConfigGenerator
The EnableFencing parameter is only used in the ConfigGenerator
to auto create templates to enable fencing. Therefore it makes
sense to have this parameter in the ConfigGenerator CRD instead
of the ControlPlane.

Also as it is a required parameter lets not specify a default and
have the user to explicitely either enable or disable it so that
the user is being forced to set either state.

(cherry picked from commit 6dc6f8b04ba6b30e021893947803b5b666e8d0fb)