Allow updating password secret with new/missing passwords
Right now the tripleo password secret only gets created once if
it does not exist. If a new operator version introduces a new
password, this would not get added to the password secret of an
existing environment.

This change validates if the password list is missing any entry
on the top level. If there is one missing it merges the new
password map[string]interface{} into the existing. It only validates
if the password list is complete, using the highest level password
keys.