[OSPL8-641/OSPK8-642] Add script to cleanup stale baremetalset compute entries
If the count of an OSBms gets increased above the number of available
BMH the OSBms will have compute nodes entries stuck with `hostRef: unassigned`,
even when reduce the spec.count again to a number that is available.

This adds a script which can be used to cleanup compute entries from
status of osbmset, osnet, osipset and osnetconfig. Also the reservations
from the spec of osnet.