Support creating bundle and index image for 16.2 and 17.0
This updates the bundle build process to support having default
images for 16.2 and 17.0. Right now tested with
scripts/build_and_push_images.sh . The bundle/index images are
now tagged with VERSION-OSP_RELEASE, e.g. 17.0.33-16.2 / 17.0.33-17.0.

In a follow up we have to update the github actions scripts to
follow the same and call the latest images latest-16.2/latest-17.0
and v1.3.x-latest-16.2 / v1.3.x-latest-17.0.