Implement ceph data export
Store the CephClientUserName from the heat environment if set,
if not then fall back to the default ("openstack").

Add a python script that mimics the python-tripleoclient export ceph logic.
Run this after every deploy to generate the ceph export yaml.
Also the merge strategy to allow multiple ceph export environment files to be combined.
If ceph is not deployed then just generate a valid empty heat environment file.
Store the ceph export yaml on the existing exported data config map.