Use sudo to run chown in init.sh of openstackclient
cephadm creates a key for the user used to deploy ceph in OSP17.
This the cloud-admin user does not have permissions to change the
ownership of this file in the next restart and the init container
fails with:

~~~
+ chown -R cloud-admin: /home/cloud-admin/.ssh
chown: changing ownership of '/home/cloud-admin/.ssh/foo': Operation not permitted
~~~

For now run the chown using sudo until we figure out other options.