Ignore unassigned hostrefs in VerifyBaremetalStatusHostRefs
If an error occurs during the baremetalset reconile loop the status can
contain incomplete hostrefs with just the initial state ("unassigned").
The hostref validation in the next reconcile then fails as it expects a
BMH with the literal name "unassigned".

This patch updates VerifyBaremetalStatusHostRefs to skip any hostref
that are still in this initial state.