Add FFU multi-rhel support
Adds tripleoRoleOverride param to config generator to allow a different t-h-t
role to be applied to a subset of hosts in heat.
Heat indices are shuffled around to compensate but the actual hostnames remain
unchanged.

For example:
    tripleoRoleOverride:
      Compute:
        roleName: ComputeRhel8
        hostIndexMap:
          1: true
          3: true
      ComputeFoo:
        roleName: ComputeFooRhel8
        hostIndexMap:
          0: true
          1: false
          2: false
          3: false
          4: false
          5: true

compute-0 remains the same in heat
compute-1 becomes computerhel8-0 in heat (still compute-1 in hostnamemap)
compute-2 becomes compute-1 in heat (still compute-2 in hostnamemap)
compute-3 becomes computerhel8-1 (still compute-3 in hostnamemap)
compute-4 becomes compute-2 in heat (still compute-4 in hostnamemap)

Then, as per a regular director-driven deployment, the role params for
the new role can be used to override the contain images etc... just for
that role to use the rhel8 config.

hostIndexMap entries default to false if not set, but can be set explicitly
e.g to keep track of which hosts have already been migrated to rhel9